Replace old DefaultLoadImageDataUseCase to new DefaultLoadImageDataUseCase
