+++ B&WTests/Domain/DefaultLoadImageDataUseCaseTests.swift
@@ -2,95 +2,79 @@
 //  DefaultLoadImageDataUseCaseTests.swift
 //  B&WTests
 //
-//  Created by Tsz-Lung on 04/04/2024.
+//  Created by Tsz-Lung on 05/04/2024.
 //
 
 import XCTest
 @testable import B_W
 
 final class DefaultLoadImageDataUseCaseTests: XCTestCase {
-    func test_init_doesNotNotifyService() {
-        let (_, service) = makeSUT()
+    func test_init_doesNotNotifyRepository() {
+        let (_, repository) = makeSUT()
         
-        XCTAssertEqual(service.requestCallCount, 0)
+        XCTAssertEqual(repository.fetchImageDataCallCount, 0)
     }
     
-    func test_load_passesCorrectParamsToService() throws {
-        let (sut, service) = makeSUT()
+    func test_load_passesURLToRepositoryCorrectly() throws {
+        let (sut, repository) = makeSUT()
         let expectedURL = URL(string: "https://image-data.com")!
         
         _ = sut.load(for: expectedURL) { _ in }
         
-        let request = try service.endpoints.first?.urlRequest()
-        XCTAssertEqual(request?.url, expectedURL)
-        XCTAssertEqual(request?.httpMethod, "GET")
+        XCTAssertEqual(repository.loggedURLs, [expectedURL])
     }
     
-    func test_load_deliversFailedErrorOnServiceError() {
-        let (sut, service) = makeSUT()
+    func test_load_deliversFailedErrorOnRepositoryError() {
+        let (sut, repository) = makeSUT()
         
         expect(sut, completeWith: .failure(.failed), when: {
-            service.complete(with: .notConnected)
+            repository.complete(with: .notConnected)
         })
     }
     
     func test_load_deliversNoDataErrorWhenReceivedNilData() {
-        let (sut, service) = makeSUT()
+        let (sut, repository) = makeSUT()
         
         expect(sut, completeWith: .failure(.noData), when: {
-            service.complete(with: nil)
+            repository.complete(with: nil)
         })
     }
     
-    func test_load_deliversDataWhenReceivedData() {
-        let (sut, service) = makeSUT()
-        let expectedData = Data("data".utf8)
+    func test_load_deliversImageDataWhenReceivedData() {
+        let (sut, repository) = makeSUT()
+        let expectedImageData = Data("image data".utf8)
         
-        expect(sut, completeWith: .success(expectedData), when: {
-            service.complete(with: expectedData)
+        expect(sut, completeWith: .success(expectedImageData), when: {
+            repository.complete(with: expectedImageData)
         })
     }
     
     func test_load_cancelsRequestSuccessfully() {
-        let (sut, service) = makeSUT()
+        let (sut, repository) = makeSUT()
         let anyData = Data("data".utf8)
         
         let task = sut.load(for: anyURL()) { _ in }
         
-        XCTAssertEqual(service.cancelCallCount, 0)
+        XCTAssertEqual(repository.cancelCallCount, 0)
         
         task.cancel()
-        service.complete(with: anyData)
+        repository.complete(with: anyData)
         
-        XCTAssertEqual(service.cancelCallCount, 1)
-    }
-    
-    func test_load_doesNotDeliverResultAfterRequestCancelled() {
-        let (sut, service) = makeSUT()
-        let anyData = Data("data".utf8)
-        
-        var completionCallCount = 0
-        let task = sut.load(for: anyURL()) { _ in completionCallCount += 1 }
-        task.cancel()
-        service.complete(with: anyData)
-        service.complete(with: nil)
-        service.complete(with: .notConnected)
-        
-        XCTAssertEqual(completionCallCount, 0)
+        XCTAssertEqual(repository.cancelCallCount, 1)
     }
     
     // MARK: - Helpers
     
     private func makeSUT(file: StaticString = #filePath,
-                         line: UInt = #line) -> (sut: LoadImageDataUseCase, service: NetworkServiceSpy) {
-        let service = NetworkServiceSpy()
-        let sut = DefaultLoadImageDataUseCase(service: service, makeRequestable: URLEndpoint.init)
-        trackForMemoryLeaks(service, file: file, line: line)
+                         line: UInt = #line) -> (sut: DefaultLoadImageDataUseCase, repository: ImageDataRepositorySpy) {
+        let repository = ImageDataRepositorySpy()
+        let sut = DefaultLoadImageDataUseCase(repository: repository)
+        trackForMemoryLeaks(repository, file: file, line: line)
         trackForMemoryLeaks(sut, file: file, line: line)
-        return (sut, service)
+        return (sut, repository)
     }
     
-    private func expect(_ sut: LoadImageDataUseCase,
+    private func expect(_ sut: DefaultLoadImageDataUseCase,
                         completeWith expectedResult: Result<Data, DefaultLoadImageDataUseCase.Error>,
                         when action: () -> Void,
                         file: StaticString = #filePath,
@@ -112,5 +96,45 @@ final class DefaultLoadImageDataUseCaseTests: XCTestCase {
         }
         action()
         wait(for: [exp], timeout: 1)
+    }
+    
+    private final class ImageDataRepositorySpy: ImageDataRepository {
+        private struct Event {
+            let url: URL
+            let completion: Completion
+        }
+        
+        private var events = [Event]()
+        var fetchImageDataCallCount: Int {
+            events.count
+        }
+        var loggedURLs: [URL] {
+            events.map(\.url)
+        }
+        
+        private struct RepositoryCancellable: Cancellable {
+            let afterCancel: () -> Void
+            
+            func cancel() {
+                afterCancel()
+            }
+        }
+        
+        private(set) var cancelCallCount = 0
+        
+        func fetchImageData(for url: URL, completion: @escaping Completion) -> Cancellable {
+            events.append(Event(url: url, completion: completion))
+            return RepositoryCancellable(afterCancel: { [weak self] in
+                self?.cancelCallCount += 1
+            })
+        }
+        
+        func complete(with error: NetworkError, at index: Int = 0) {
+            events[index].completion(.failure(error))
+        }
+        
+        func complete(with data: Data?, at index: Int = 0) {
+            events[index].completion(.success(data))
+        }
     }
 }
